@@ -1,49 +1,50 @@
 {
-  "name": "angular-rtc",
-  "version": "0.0.0",
-  "license": "MIT",
-  "scripts": {
-    "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
-  },
-  "private": true,
-  "dependencies": {
-    "@angular/animations": "^5.0.0",
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/forms": "^5.0.0",
-    "@angular/http": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/router": "^5.0.0",
-    "core-js": "^2.4.1",
-    "rxjs": "^5.5.2",
-    "zone.js": "^0.8.14"
-  },
-  "devDependencies": {
-    "@angular/cli": "^1.6.0",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/language-service": "^5.0.0",
-    "@types/jasmine": "~2.5.53",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "~3.0.1",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~3.0.4",
-    "tslint": "~5.3.2",
-    "typescript": "~2.4.2"
-  }
+	"name": "angular-rtc",
+	"version": "0.0.0",
+	"license": "MIT",
+	"scripts": {
+		"ng": "ng",
+		"start": "ng serve",
+		"build": "ng build",
+		"test": "ng test",
+		"lint": "ng lint",
+		"e2e": "ng e2e"
+	},
+	"private": true,
+	"dependencies": {
+		"@angular/animations": "^6.0.0",
+		"@angular/common": "^6.0.0",
+		"@angular/compiler": "^6.0.0",
+		"@angular/core": "^6.0.0",
+		"@angular/forms": "^6.0.0",
+		"@angular/http": "^6.0.0",
+		"@angular/platform-browser": "^6.0.0",
+		"@angular/platform-browser-dynamic": "^6.0.0",
+		"@angular/router": "^6.0.0",
+		"core-js": "^2.4.1",
+		"rxjs": "^6.2.2",
+		"zone.js": "^0.8.26"
+	},
+	"devDependencies": {
+		"@angular/cli": "^6.0.0",
+		"@angular/compiler-cli": "^6.0.0",
+		"@angular/language-service": "^6.0.0",
+		"@types/jasmine": "~2.5.53",
+		"@types/jasminewd2": "~2.0.2",
+		"@types/node": "~6.0.60",
+		"codelyzer": "~4.3.0",
+		"jasmine-core": "~2.6.2",
+		"jasmine-spec-reporter": "~4.1.0",
+		"karma": "~1.7.0",
+		"karma-chrome-launcher": "~2.1.1",
+		"karma-cli": "~1.0.1",
+		"karma-coverage-istanbul-reporter": "^1.2.1",
+		"karma-jasmine": "~1.1.0",
+		"karma-jasmine-html-reporter": "^0.2.2",
+		"protractor": "~5.1.2",
+		"ts-node": "~3.0.4",
+		"tslint": "~5.3.2",
+		"typescript": "~2.7.2",
+		"@angular-devkit/build-angular": "~0.6.8"
+	}
 }
